deplace api-p support
