+++ src/ApiPlatform/State/RedisProvider.php
@@ -2,11 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Talleu\RedisOm\Bundle\ApiPlatform\State;
+namespace Talleu\RedisOm\ApiPlatform\State;
 
 use ApiPlatform\Metadata\CollectionOperationInterface;
 use ApiPlatform\Metadata\Operation;
-use ApiPlatform\Metadata\Operations;
 use ApiPlatform\State\ProviderInterface;
 
 class RedisProvider implements ProviderInterface
