+++ src/Bundle/Resources/config/api_platform.xml
@@ -5,48 +5,48 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="talleu_php_redis_om.api_platform.query_extension.filters" class="Talleu\RedisOm\Bundle\ApiPlatform\Extensions\FilterExtension" public="false">
+        <service id="talleu_php_redis_om.api_platform.query_extension.filters" class="Talleu\RedisOm\ApiPlatform\Extensions\FilterExtension" public="false">
             <tag name="talleu_php_redis_om.api_platform.query_extension.collection" priority="-64" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.query_extension.pagination" class="Talleu\RedisOm\Bundle\ApiPlatform\Extensions\PaginationExtension" public="false">
+        <service id="talleu_php_redis_om.api_platform.query_extension.pagination" class="Talleu\RedisOm\ApiPlatform\Extensions\PaginationExtension" public="false">
             <argument type="service" id="talleu_php_redis_om.object_manager" />
             <argument type="service" id="api_platform.pagination" />
 
             <tag name="talleu_php_redis_om.api_platform.query_extension.collection" priority="-64" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.state.collection_provider" class="Talleu\RedisOm\Bundle\ApiPlatform\State\CollectionProvider" public="false">
+        <service id="talleu_php_redis_om.api_platform.state.collection_provider" class="Talleu\RedisOm\ApiPlatform\State\CollectionProvider" public="false">
             <argument type="service" id="talleu_php_redis_om.object_manager" />
             <argument type="tagged_iterator" tag="talleu_php_redis_om.api_platform.query_extension.collection" />
 
-            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\Bundle\ApiPlatform\State\CollectionProvider" />
+            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\ApiPlatform\State\CollectionProvider" />
             <tag name="api_platform.state_provider" priority="-100" key="talleu_php_redis_om.api_platform.state.collection_provider" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.state.processor" class="Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProcessor" public="false">
+        <service id="talleu_php_redis_om.api_platform.state.processor" class="Talleu\RedisOm\ApiPlatform\State\RedisProcessor" public="false">
             <argument type="service" id="talleu_php_redis_om.object_manager" />
 
-            <tag name="api_platform.state_processor" priority="-100" key="Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProcessor" />
+            <tag name="api_platform.state_processor" priority="-100" key="Talleu\RedisOm\ApiPlatform\State\RedisProcessor" />
             <tag name="api_platform.state_processor" priority="-100" key="talleu_php_redis_om.api_platform.state.processor" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.state.item_provider" class="Talleu\RedisOm\Bundle\ApiPlatform\State\ItemProvider" public="false">
+        <service id="talleu_php_redis_om.api_platform.state.item_provider" class="Talleu\RedisOm\ApiPlatform\State\ItemProvider" public="false">
             <argument type="service" id="talleu_php_redis_om.object_manager" />
 
-            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\Bundle\ApiPlatform\State\ItemProvider" />
+            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\ApiPlatform\State\ItemProvider" />
             <tag name="api_platform.state_provider" priority="-100" key="talleu_php_redis_om.api_platform.state.item_provider" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.state.redis_provider" class="Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProvider" public="false">
+        <service id="talleu_php_redis_om.api_platform.state.redis_provider" class="Talleu\RedisOm\ApiPlatform\State\RedisProvider" public="false">
             <argument type="service" id="talleu_php_redis_om.api_platform.state.collection_provider" />
             <argument type="service" id="talleu_php_redis_om.api_platform.state.item_provider" />
 
-            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProvider" />
+            <tag name="api_platform.state_provider" priority="-100" key="Talleu\RedisOm\ApiPlatform\State\RedisProvider" />
             <tag name="api_platform.state_provider" priority="-100" key="talleu_php_redis_om.api_platform.state.redis_provider" />
         </service>
 
-        <service id="talleu_php_redis_om.api_platform.metadata.resource.metadata_collection_factory" class="Talleu\RedisOm\Bundle\ApiPlatform\Metadata\ResourceCollectionMetadataFactory" decorates="api_platform.metadata.resource.metadata_collection_factory" decoration-priority="40">
+        <service id="talleu_php_redis_om.api_platform.metadata.resource.metadata_collection_factory" class="Talleu\RedisOm\ApiPlatform\Metadata\ResourceCollectionMetadataFactory" decorates="api_platform.metadata.resource.metadata_collection_factory" decoration-priority="40">
             <argument type="service" id="talleu_php_redis_om.api_platform.metadata.resource.metadata_collection_factory.inner" />
         </service>
     </services>
