+++ tests/ApiPlatform/Entity/Dummy.php
@@ -4,16 +4,15 @@ declare(strict_types=1);
 
 namespace Talleu\RedisOm\Tests\ApiPlatform\Entity;
 
-use ApiPlatform\Laravel\Eloquent\Filter\JsonApi\SortFilter;
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\QueryParameter;
-use Talleu\RedisOm\Bundle\ApiPlatform\Filters\BooleanFilter;
-use Talleu\RedisOm\Bundle\ApiPlatform\Filters\ExactSearchFilter;
-use Talleu\RedisOm\Bundle\ApiPlatform\Filters\NumericFilter;
-use Talleu\RedisOm\Bundle\ApiPlatform\Filters\OrderFilter;
-use Talleu\RedisOm\Bundle\ApiPlatform\Filters\SearchFilter;
-use Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProcessor;
-use Talleu\RedisOm\Bundle\ApiPlatform\State\RedisProvider;
+use Talleu\RedisOm\ApiPlatform\Filters\BooleanFilter;
+use Talleu\RedisOm\ApiPlatform\Filters\ExactSearchFilter;
+use Talleu\RedisOm\ApiPlatform\Filters\NumericFilter;
+use Talleu\RedisOm\ApiPlatform\Filters\OrderFilter;
+use Talleu\RedisOm\ApiPlatform\Filters\SearchFilter;
+use Talleu\RedisOm\ApiPlatform\State\RedisProcessor;
+use Talleu\RedisOm\ApiPlatform\State\RedisProvider;
 use Talleu\RedisOm\Om\Mapping as RedisOm;
 use Talleu\RedisOm\Tests\Fixtures\Hash\DummyHash;
